+++ examples/eu-west-1-prod/main.tf
@@ -5,7 +5,9 @@ module "aws-ecs-app" {
   service_name      = var.service_name
   environment       = var.environment
   cdn_secret_header = var.cdn_secret_header
+  app_secret_parameter = var.app_secret_parameter
   zone_id           = var.zone_id
+  app_image         = var.app_image
 }
 
 output "app_https_url" {