+++ examples/eu-west-1-prod/variables.tf
@@ -27,6 +27,12 @@ variable "app_fargate_memory" {
   default     = "512"
 }
 
+variable "app_secret_parameter" {
+  description = "The App secret parameter"
+  type        = string
+  sensitive   = true
+}
+
 variable "cdn_secret_header" {
   description = "Secret header for CDN"
   type        = string
@@ -36,4 +42,10 @@ variable "cdn_secret_header" {
 variable "zone_id" {
   description = "Route53 zone ID"
   type        = string
+}
+
+variable "app_image" {
+  description = "The CIDR block for the VPC"
+  type        = string
+  default     = "cillu/nodejs-app:1.0.0"
 }